chore: refactor queries
